Update gotcha to accept configuration via ENV vars
`gotcha` can now run as a Cloud Foundry app. It pulls
`PORT` from the environment to determine what to bind to,
and its target from the `GOTCHA_BACKEND` environment variable.

The `PORT` env var is only used if no ports are explicitly
configured via the CLI invocation of `gotcha`. If neither
CLI args nor `PORT` env var exists, fails back to port 3128.

The `GOTCHA_BACKEND` env var is only used when there are no
CLI args to gotcha at all. Otherwise, the behavior is the same
as it previously was.

Additionally, the `SSL_SKIP_VERIFY` environment variable
can be used to disable SSL certificate checking of the
backend gotcha is hooked up to.

And for kicks, the HTTP_PROXY/HTTPS_PROXY/NO_PROXY env
vars are now also supported